Move FileDialogStorage to config module
