+++ src/lib.rs
@@ -162,7 +162,7 @@ mod file_dialog;
 mod modals;
 
 pub use config::{
-    FileDialogConfig, FileDialogKeyBindings, FileDialogLabels, IconFilter, KeyBinding, QuickAccess,
-    QuickAccessPath,
+    FileDialogConfig, FileDialogKeyBindings, FileDialogLabels, FileDialogStorage, IconFilter,
+    KeyBinding, QuickAccess, QuickAccessPath,
 };
-pub use file_dialog::{DialogMode, DialogState, FileDialog, FileDialogStorage};
+pub use file_dialog::{DialogMode, DialogState, FileDialog};
